requirements.txt
FastAPI
asyncpg
aiohttp
pyyaml
aiosmtplib
bcrypt
snowflake-id